# customers-api/openapi.yaml
openapi: 3.0.3
info:
  title: Customers API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
...
  /auth/login:
    post:
      summary: Login simple para obtener JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Token generado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceTokenAuth:
      type: http
      scheme: bearer